Finish playRound() function
Finishes the playRound() function. In this funcion
we take playerSelection and computerSelection as
function parameters.
We convert both of those to lower case to make the
parameters case insensitive.
Used if, else if and else statements to figure out
if who wins. If the user wins then a boolean value
didPlayerWin will be set to true.
didPlayerWin value is finally returned.